image_editing/requirements.txt
Django
djangorestframework
uvicorn
python-multipart
ffmpeg-python
